DjAssignment/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.3
